requirements.txt
openai
SpeechRecognition
pyaudio
pyttsx3
python-dotenv
pygame
gTTS
gradio